* implement slurm epilog and slurm prolog handling to allow correct job synchronization
  * fixing job timout bug
  * move removement of toml files to backend
  * rewrite jobtable
  * rewrite job clean up
  * remove DB entries (no matter what the respective exit code is)
* follow the _XDG Base Directory Specifications_ for folders created in `${HOME}`
* minor bug fixes
* redefine `mpirun` command to work smoothly inside the containers
* removes some previously hardcoded stuff and replaces it with dynamic variables
* reduces the user bashrc to the minimum needed and puts everything else in
  separate files
* restricts the sourced variables in all scripts to those that are really needed
* updates the scripts that start/stop the proxy and frontend containers
* the script that splits the config file into separate files is updated and takes
  care of backups

+++ Carme-Scripts/slurm/slurm-epilog-scripts/carme-slurm-clean.sh
@@ -31,38 +31,32 @@ fi
 #-----------------------------------------------------------------------------------------------------------------------------------
 
 
-#-----------------------------------------------------------------------------------------------------------------------------------
-# remove stuff in ${HOME} ----------------------------------------------------------------------------------------------------------
+# delte CARME specific job folders -------------------------------------------------------------------------------------------------
+# delete local scratch folder ------------------------------------------------------------------------------------------------------
+if [[ -d "/scratch_local/${SLURM_JOB_ID}" ]];then
+  rm -r /scratch_local/${SLURM_JOB_ID} 
+fi
 
+
+# remove stuff in ${HOME} ----------------------------------------------------------------------------------------------------------
+MASTER_NODE=$(scontrol show job ${SLURM_JOB_ID} | grep BatchHost | cut -d "=" -f 2 | cut -d/ -f1)
 USER_HOME=$(getent passwd ${SLURM_JOB_USER} | cut -d: -f6)
 
-# delete .bash_carme_$SLURM_JOB_ID
-if [[ -f "${USER_HOME}/.carme/.bash_carme_${SLURM_JOB_ID}" ]];then
-  rm ${USER_HOME}/.carme/.bash_carme_${SLURM_JOB_ID}
-fi
+if [[ "$(hostname)" == "${MASTER_NODE}" ]];then
+  # remove job specific stuff
+  if [[ -d "${USER_HOME}/.local/share/carme/tmp-files-${SLURM_JOB_ID}" ]];then
+    rm -r ${USER_HOME}/.local/share/carme/tmp-files-${SLURM_JOB_ID}
+  fi
 
-# delete local scratch folder
-if [[ -d "/scratch_local/${SLURM_JOB_ID}" ]];then
-  rm -r /scratch_local/${SLURM_JOB_ID}
-fi
+  # remove job tensorboard folder
+  if [[ -d "${USER_HOME}/tensorboard/tensorboard_${SLURM_JOB_ID}" ]];then
+    rm -r ${USER_HOME}/tensorboard/tensorboard_${SLURM_JOB_ID}
+  fi
 
-# remove job tensorboard folder
-if [[ -d "${USER_HOME}/tensorboard/tensorboard_${SLURM_JOB_ID}" ]];then
-  rm -r ${USER_HOME}/tensorboard/tensorboard_${SLURM_JOB_ID}
-fi
-
-# remove theia tmp folder
-if [[ -d "${USER_HOME}/carme_tmp/${SLURM_JOB_ID}_job_tmp" ]];then
-  rm -r ${USER_HOME}/carme_tmp/${SLURM_JOB_ID}_job_tmp
-fi
-
-## remove job ssh stuff
-if [[ -f "${USER_HOME}/.ssh/id_rsa_${SLURM_JOB_ID}" ]];then
-  rm ${USER_HOME}/.ssh/id_rsa_${SLURM_JOB_ID}
-fi
-
-if [[ -d "${USER_HOME}/.carme/tmp_ssh_${SLURM_JOB_ID}" ]];then
-  rm -r ${USER_HOME}/.carme/tmp_ssh_${SLURM_JOB_ID}
+  # remove job ssh stuff
+  if [[ -f "${USER_HOME}/.ssh/id_rsa_${SLURM_JOB_ID}" ]];then
+    rm ${USER_HOME}/.ssh/id_rsa_${SLURM_JOB_ID}
+  fi
 fi
 #-----------------------------------------------------------------------------------------------------------------------------------
